fix circular references
