+++ packages/react/runtime/src/pendingListUpdates.ts
@@ -1,0 +1,18 @@
+// Copyright 2025 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
+
+import type { ListUpdateInfo } from './listUpdateInfo.js';
+
+export const __pendingListUpdates = {
+  values: {} as Record<number, ListUpdateInfo>,
+  clear(): void {
+    this.values = {};
+  },
+  flush(): void {
+    Object.values(this.values).forEach(update => {
+      update.flush();
+    });
+    this.clear();
+  },
+};
